Extend fetch commit timeout for cold sync
